requirements.txt
flask-sqlalchemy
flask-login
flask-bcrypt
email_validator
